+++ src/Data/Geo/Jord/LatLong.hs
@@ -11,7 +11,9 @@
 module Data.Geo.Jord.LatLong
     (
     -- * The 'LatLong' type
-      LatLong(latitude, longitude)
+      LatLong
+    , latitude
+    , longitude
     -- * Smart constructors
     , latLong
     , latLongE
@@ -39,8 +41,8 @@ import Text.Read hiding (pfail)
 
 -- | Horizontal position defined by its geodetic latitude and longitude.
 data LatLong = LatLong
-    { latitude :: Angle
-    , longitude :: Angle
+    { latitude :: Angle -- ^ geodetic latitude
+    , longitude :: Angle -- ^ longitude
     } deriving (Eq)
 
 -- | See 'readLatLong'.