+++ ChangeLog.md
@@ -1,3 +1,10 @@
+### 0.4.1.0
+
+- Fixed interceptBySpeed
+- Nautical miles symbol is "nm"
+- REPL: intercept for intercept, interceptBySpeed and interceptByTime
+- REPL: show length and speed in user selected unit
+
 ### 0.4.0.0
 
 - Added ECEF, frames and delta to REPL